Process existing assets in batch
Fix for https://github.com/google/github-release-retry/issues/4
A Release fetched though the GitHub API v3 already contains all existing
Asset details (in particular the status and size). We take advantage of
this to avoid doing extra API checks for each local asset if a local
asset is found to already exist remotely in the GitHub release that was
already fetched. Assets that are not already uploaded are then processed
normally for direct upload without the need to further check that they
need uploading since we already know they are missing.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>